Добавляет компонент favorites-empty
